fix: Fix matching bonus number error
보너스 번호를 LottoTicket에서 LottoDiscriminator로 옮김.
LottoNumbers를 지우고, 그 책임을 LottoTicket으로 합침.
잘못된 설계로 보너스 번호 체크가 제대로 안 됐던 에러를 설계를 바꾸면서
고쳐짐.
