+++ src/main/java/lotto/domain/LottoTicket.java
@@ -1,36 +1,47 @@
 package lotto.domain;
 
+import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
 
 public class LottoTicket {
-    private static final String BONUS_NUMBER_DUPLICATE_ERROR_MESSAGE = "보너스 번호는 다른 로또 번호와 중복될 수 없습니다.";
+    public static final int LOTTO_NUMBERS_SIZE = 6;
+    private static final String SIZE_ERROR_MESSAGE = "로또 번호는 6개여야 합니다.";
+    private static final String DUPLICATE_ERROR_MESSAGE = "로또 번호는 중복될 수 없습니다.";
 
-    private final LottoNumbers lottoNumbers;
-    private final LottoNumber bonusNumber;
+    private final List<LottoNumber> lottoNumbers;
 
-    public LottoTicket(LottoNumbers lottoNumbers, LottoNumber bonusNumber) {
-        validateDuplicateBonusNumber(lottoNumbers, bonusNumber);
+    public LottoTicket(List<LottoNumber> lottoNumbers) {
+        validateSize(lottoNumbers);
+        validateDuplicate(lottoNumbers);
 
         this.lottoNumbers = lottoNumbers;
-        this.bonusNumber = bonusNumber;
     }
 
-    private void validateDuplicateBonusNumber(LottoNumbers lottoNumbers, LottoNumber bonusNumber) {
-        if (lottoNumbers.contains(bonusNumber)) {
-            throw new IllegalArgumentException(BONUS_NUMBER_DUPLICATE_ERROR_MESSAGE);
+    private void validateSize(List<LottoNumber> lottoNumbers) {
+        if (lottoNumbers.size() != LOTTO_NUMBERS_SIZE) {
+            throw new IllegalArgumentException(SIZE_ERROR_MESSAGE);
         }
     }
 
-    public LottoNumbers lottoNumbers() {
-        return lottoNumbers;
+    private void validateDuplicate(List<LottoNumber> lottoNumbers) {
+        if (lottoNumbers.stream().distinct().count() != lottoNumbers.size()) {
+            throw new IllegalArgumentException(DUPLICATE_ERROR_MESSAGE);
+        }
     }
 
-    public int matchingLottoNumbersCount(LottoTicket that) {
-        return lottoNumbers.matchingCount(that.lottoNumbers);
+    public int matchingCount(LottoTicket that) {
+        return (int) lottoNumbers.stream()
+                .filter(that::contains)
+                .count();
     }
 
-    public boolean sameBonusNumber(LottoTicket that) {
-        return bonusNumber == that.bonusNumber;
+    public boolean contains(LottoNumber lottoNumber) {
+        return lottoNumbers.contains(lottoNumber);
+    }
+
+    public List<LottoNumber> lottoNumbers() {
+        return Collections.unmodifiableList(lottoNumbers);
     }
 
     @Override
@@ -38,11 +49,11 @@ public class LottoTicket {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         LottoTicket that = (LottoTicket) o;
-        return Objects.equals(lottoNumbers, that.lottoNumbers) && Objects.equals(bonusNumber, that.bonusNumber);
+        return Objects.equals(lottoNumbers, that.lottoNumbers);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(lottoNumbers, bonusNumber);
+        return Objects.hash(lottoNumbers);
     }
 }
